pyproject.toml
[project]
name = "avl-test-setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "avl-core>=0.5.0",
    "cocotb>=2.0.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
